# internal/adapter/docsui/openapi/tl.yaml
openapi: 3.0.3
info:
  title: ANS Transparency Log API
  version: "2.0.0"
  description: |
    OpenAPI contract for the ANS Transparency Log (`ans-tl`). Paths
    match the reference TL's swagger byte-for-byte — `/v1/agents/*`,
    `/v1/log/*`, `/v1/internal/*`, C2SP tlog-tiles at root
    (`/checkpoint`, `/root-keys`, `/tile/*`), and
    `/internal/v1/producer-keys/*` for admin CRUD.

    Behavioral parity with the reference TL:

      - Event envelope is the V2 shape — nested `{payload:{logId,
        producer:{event, keyId, signature}}, schemaVersion:"V2",
        signature, status}`. V2 evolves the reference V1 with
        unified `identityCerts[]` / `serverCerts[]` arrays, typed
        `dnsRecordsProvisioned[]` records, and `SHA256:`-prefixed
        fingerprints. V0 + V1 schemas are served for historical
        verifier backwards compatibility (never emitted).
      - Receipts are SCITT COSE_Sign1 (RFC 8152 tag 18) with VDS 1
        (RFC 9162_SHA256) and the VDP inclusion proof in the
        unprotected header (label 396). See
        `internal/tl/receipt/constants.go` for the concrete labels.
      - `/root-keys` emits the sumdb-note verification format
        byte-for-byte compatible with the reference.
      - `/internal/v1/producer-keys` admin routes match the reference
        swagger.yaml snake_case DTOs; the only ans-only extension is
        `GET /internal/v1/producer-keys/{key_id}` (reference doesn't
        expose single-record lookup).
servers:
...

...
security:
  - bearerAuth: []
paths:
...
  /internal/v1/producer-keys:
    post:
      tags: [Admin]
      summary: Register a new producer key
      operationId: createProducerKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerKeyRequest'
      responses:
        '200':
          description: Key registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProducerKeyResponse'
        '401': { description: Unauthorized }
        '403': { description: Admin privileges required }
        '409':
          description: Producer key with this key_id already exists.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: |
            Validation failed. Codes: `INVALID_DATE_RANGE`,
            `INVALID_PUBLIC_KEY_PEM`, `BAD_JSON`, `BAD_BODY`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /internal/v1/producer-keys/ra/{ra_id}:
    get:
      tags: [Admin]
      summary: List producer keys by RA
      operationId: listProducerKeysByRaId
      parameters:
        - name: ra_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Keys for the RA (includes revoked rows for audit).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProducerKeysListResponse'
        '401': { description: Unauthorized }
        '403': { description: Admin privileges required }
        '404':
          description: No producer keys found for this RA ID.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /internal/v1/producer-keys/{key_id}:
    get:
      tags: [Admin]
      summary: Fetch a single producer key by ID (ans-only extension)
      description: |
        Not in the reference swagger. Added as a convenience for admin
        CLIs that already hold a key_id (e.g., after a successful
        create) and want to read it back without listing-by-ra.
      operationId: getProducerKey
      parameters:
        - name: key_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Key record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProducerKeyResponse'
        '401': { description: Unauthorized }
        '403': { description: Admin privileges required }
        '404':
          description: No such key_id.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags: [Admin]
      summary: Revoke a producer key
      operationId: deleteProducerKey
      parameters:
        - name: key_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Key revoked.
        '401': { description: Unauthorized }
        '403': { description: Admin privileges required }
        '404':
          description: Key not found or already revoked.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  # ──────────────────────────────────────────────────────────────
  # Health
  # ──────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static API key or OIDC bearer token. Admin routes
        (`/internal/v1/producer-keys/*`) also require the caller to
        be an admin — static keys are always admin in the
        quickstart; OIDC tokens inherit admin-ness from the
        configured admin-groups list. Verifier GETs are anonymous
        when `cfg.auth.public-read: true`.
...
  schemas:
...
    ProducerKeyRequest:
      type: object
      required: [key_id, public_key_pem, algorithm, ra_id, valid_from, expires_at]
      properties:
        key_id:         { type: string }
        public_key_pem: { type: string, description: "PEM-wrapped SPKI" }
        algorithm:      { type: string, enum: [ES256] }
        ra_id:          { type: string }
        valid_from:     { type: string, format: date-time }
        expires_at:     { type: string, format: date-time }
        metadata:
          $ref: '#/components/schemas/ProducerKeyMetadata'
    ProducerKeyResponse:
      type: object
      required: [key_id, status, fingerprint, created_at]
      properties:
        key_id:         { type: string }
        key_id_opaque:  { type: string, nullable: true }
        ra_id_opaque:   { type: string, nullable: true }
        status:         { type: string, enum: [active, revoked] }
        fingerprint:    { type: string, example: "SHA256:<hex>" }
        created_at:     { type: string, format: date-time }
        # Fields repeated on the full record (returned by GET-by-id
        # + present on list items — matches reference §1481).
        public_key_pem: { type: string }
        algorithm:      { type: string }
        ra_id:          { type: string }
        valid_from:     { type: string, format: date-time }
        expires_at:     { type: string, format: date-time }
        revoked_at:     { type: string, format: date-time, nullable: true }
        metadata:
          $ref: '#/components/schemas/ProducerKeyMetadata'
    ProducerKeyMetadata:
      type: object
      description: Free-form JSON; reference uses environment/region/rotation_of.
      additionalProperties: true
      properties:
        environment: { type: string, enum: [production, stage, development, testing] }
        region:      { type: string }
        rotation_of: { type: string, nullable: true }
    ProducerKeysListResponse:
      type: object
      required: [keys, total_count]
      properties:
        keys:
          type: array
          items:
            $ref: '#/components/schemas/ProducerKeyResponse'
        total_count:
          type: integer
          format: int64

    # ── Checkpoint JSON shapes ────────────────────────────────
...
    Problem:
      type: object
      description: RFC 7807 problem details.
      properties:
        type:   { type: string, example: "about:blank" }
        title:  { type: string }
        status: { type: integer }
        code:   { type: string, pattern: "^[A-Z_][A-Z0-9_]*$" }
        detail: { type: string }